＃89 PR review respond refactor: remove legacy IOAPIC init and PciAllocator, expose PCI raw write
- Remove configure_io_apic_for_legacy_irqs and init_io_apic functions (unused).
- Delete PciAllocator struct and assign_bars method; adjust init.rs to pass devices directly.
- Make write_config_dword_raw public with proper documentation.
- Move enable_memory_access from constructor to separate public method for
