+++ fullerene-kernel/src/init.rs
@@ -64,7 +64,7 @@ pub fn init_common(physical_memory_offset: x86_64::VirtAddr) {
             let mut scanner = petroleum::hardware::pci::PciScanner::new();
             if scanner.scan_all_buses().is_ok() {
                 let mut allocator = petroleum::hardware::pci::PciAllocator::new(0x40000000);
-                allocator.assign_bars(&scanner);
+                allocator.assign_bars(scanner.get_devices());
             }
             Ok(())
         }),
